% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagyo.R
\docType{package}
\name{bagyo}
\alias{bagyo-package}
\alias{bagyo}
\title{Philippine Tropical Cyclones Data}
\description{
The Philippines frequently experiences tropical cyclones (called bagyo in the
Filipino language) because of its geographical position. These cyclones
typically bring heavy rainfall, leading to widespread flooding, as well as
strong winds that cause significant damage to human life, crops, and
property. Data on cyclones are collected and curated by the Philippine
Atmospheric, Geophysical, and Astronomical Services Administration (PAGASA)
and made available through its website. This package contains Philippine
Tropical Cyclone data in a machine-readable format. It is hoped that this
data package provides an interesting and unique dataset for data exploration
and visualisation as an adjunct to the traditional \code{iris} dataset and to the
current \code{palmerpenguins} dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://panukatan.io/bagyo/}
  \item \url{https://github.com/panukatan/bagyo}
  \item Report bugs at \url{https://github.com/panukatan/bagyo/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
