% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyclones}
\alias{cyclones}
\title{Table of tropical cyclones that entered the Philippines from 2017 to 2020s}
\format{
A data frame with 9 columns and 86 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{year} \tab Year \cr
   \emph{category_code} \tab Tropical cyclone category code \cr
   \emph{category_name} \tab Tropical cyclone category name \cr
   \emph{name} \tab Name given to the tropical cyclone by Philippine authorities \cr
   \emph{rsmc_name} \tab Name given to the tropical cyclone by RSMC \cr
   \emph{start} \tab Date and time at which cyclone enters Philippine waters \cr
   \emph{end} \tab Date and time at which cyclone leaves Philippine waters \cr
   \emph{pressure} \tab Maximum central pressure in \emph{hPa} \cr
   \emph{speed} \tab Maximum sustained wind speed in \emph{km/h} \cr
}
}
\source{
Data are drawn from PAGASA's Annual Report on Philippine Tropical
Cyclones found at
https://www.pagasa.dost.gov.ph/tropical-cyclone/publications/annual-report
}
\usage{
cyclones
}
\description{
Table of tropical cyclones that entered the Philippines from 2017 to 2020s
}
\examples{
cyclones

}
\keyword{datasets}
