% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cyclone.R
\name{find_bagyo}
\alias{find_bagyo}
\title{Find cyclones information}
\usage{
find_bagyo(.year = NULL, .category = NULL)
}
\arguments{
\item{.year}{An integer value for a year or a vector of years of cyclones
data to retrieve. Default to NULL to retrieve all years.}

\item{.category}{A character value or a vector of category code/s or category
name/s to retrieve. Default to NULL to retrieve all categories}
}
\value{
A data.frame of cyclones information based on specified \code{year} and
\code{category}
}
\description{
Find cyclones information
}
\examples{
find_bagyo()
find_bagyo(.year = 2017)
find_bagyo(.category = "TD")
find_bagyo(.year = 2017, .category = "TD")

}
