% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cyclone.R
\name{get_bagyo}
\alias{get_bagyo}
\title{Get a cyclone name}
\usage{
get_bagyo()
}
\value{
A list of information regarding a randomly selected cyclone from
the \code{bagyo} dataset.
}
\description{
Randomly returns a cyclone name and prints additional information available
from the package regarding the cyclone.
}
\examples{
get_bagyo()

}
