% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cyclone.R
\name{print.bagyo}
\alias{print.bagyo}
\title{\code{\link[=print]{print()}} helper function for \code{\link[=get_bagyo]{get_bagyo()}} function}
\usage{
\method{print}{bagyo}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bagyo} produced by \code{\link[=get_bagyo]{get_bagyo()}} function}

\item{...}{Additional \code{\link[=print]{print()}} arguments}
}
\value{
Printed output of \code{\link[=get_bagyo]{get_bagyo()}} function
}
\description{
\code{\link[=print]{print()}} helper function for \code{\link[=get_bagyo]{get_bagyo()}} function
}
